Say more in the typed model than "transcode this"
Fourteen things the API could not express, and three bugs that only
surfaced once something asked for them.

**Encoding.** Constrained rate control (maxrate/bufsize, emitted only
together because x264 ignores a ceiling with no window), keyframe
interval, profile, level, tune, and an `encoderOptions` bag for the
private parameter blobs. `+faststart`, for containers where it means
something. Colour and HDR tags, settable outright or carried off a probed
stream - ffmpeg keeps them on some paths and drops them on others, and
the failure is a washed-out picture with nothing in the log.

**Two-pass is a flag.** `isTwoPass: true` and `convert` runs ffmpeg
twice, keeps the statistics between the runs and reports one continuous
progress across both. Same machinery normalises loudness: `loudnorm`
measures the whole track, then applies what it found. A conversion
wanting both pays for one extra decode, not two.

**Streams.** Per-stream overrides, so two audio tracks need not be the
same bitrate. Subtitles as a concern of their own. Channel layout, which
is not channel count. Chapters, which `ProbeOptions.showChapters` has
been asking for into a type that had nowhere to put them.

**Sources and shapes.** Remote URLs, as a second kind of `Input` -
nothing to open, so ffmpeg opens it. Joining, thumbnails and contact
sheets. And an ETA, which needed `speed` fixing first: it was parsed
with `Int()` from strings like "0.997x", so every fractional speed - very
nearly all of them - was discarded.

**Hardware stays opt-in**, as asked. Nothing picks a VideoToolbox encoder
for you, because at a given bitrate it is usually worse than x264 and
which of speed or quality matters is not this package's call.
`hardwareEncoder(for:)` answers the awkward half - whether one exists -
and `.h264VideoToolbox` / `.hevcVideoToolbox` are named explicitly.
Decoding is per-input in the same way.

**What the descriptor design costs, found by trying.**

`-passlogfile` has no descriptor form: ffmpeg opens it by name. Since
FFmpegTask inherits the service's sandbox, the service is the process
that can name somewhere both may write, so requests may now carry a
scratch token that it fills in with a path it chose. A two-pass encode is
two jobs sharing one log, so both name the same token and only the last
asks for cleanup - except on failure, where a half-written log is worse
than none.

`image2` reads its output as a filename pattern; handed `fd:` it wrote
the picture to stdout and exited 0, leaving the destination empty. Stills
mux as `image2pipe`. The same reasoning rules out numbered sequences, so
`contactSheet` tiles into one image. And joining uses the concat filter
rather than the demuxer, whose list file names its inputs by path.

**Three bugs.** `isEmptyJSON` kept only braces and checked the first two
were "{}", which is true of any flat object whose values contain no
braces - so `{"license": "..."}` was discarded as empty and the caller
was told the job had merely succeeded. It has its own tests now.
`speed` as above. And loudnorm prints at info level while FFmpegTask runs
ffmpeg at warning, so the measuring pass ran and reported nothing; the
measuring pass raises the level, and only it does.

The harness was generating a clip with no soundtrack, which is why none
of that surfaced earlier: every audio assertion in it passed vacuously.
It now generates audio and checks for it first.

Verified: FFmpegTaskTests 33, XPCFFmpegServiceTests 53, XPCFFmpeg 139,
and 59 end-to-end checks against the real service, sandboxed and
unsandboxed, all passing.

+++ FFmpegTask/OutputHandlers.swift
@@ -151,7 +151,7 @@ struct FFmpegProgress: FFmpegOutputHandlerWithTerminators {
         var outTime: TimeInterval?
         var duplicateFrames: Int
         var droppedFrames: Int
-        var speed: Int?
+        var speed: Double?
         var finished: Bool
     }
 
@@ -193,7 +193,7 @@ struct FFmpegProgress: FFmpegOutputHandlerWithTerminators {
         let bitrate = m.value(0, "Bitrate").flatMap { Double($0) }
         let totalSize = m.value(0, "TotalSize").flatMap { Int($0) }
         let ms = m.value(0, "OutTime").flatMap { TimeInterval($0) }
-        let speed = m.value(0, "Speed").flatMap { Int($0) }
+        let speed = m.value(0, "Speed").flatMap { Double($0) }
         
         self.progress = FFmpegProgressStruct(frame: frame, fps: fps, input: input, stream: stream, quality: quality, bitrate: bitrate, totalSize: totalSize, outTime: ms == nil ? nil : ms! / 1000000.0, duplicateFrames: duplicateFrames, droppedFrames: droppedFrames, speed: speed, finished: m.value(0, "Progress") == "end")
     }
